refactored components to account for new additions
