+++ client/src/components/App.jsx
@@ -1,7 +1,7 @@
 import React, { useState, useEffect } from 'react';
 import QuestionAnswerContainer from './questionsAnswers/QuestionAnswerContainer.jsx';
 import RelatedProducts from './relatedProducts/relatedProducts.jsx';
-import RatingReviews from './RatingReviews/ReviewsList.jsx';
+import RatingReviews from './RatingReviews/RatingsReviews.jsx';
 import ProductDetails from './productDetails/index.jsx';
 
 
@@ -21,7 +21,7 @@ const App = () => {
       </h1>
       {/* <ProductDetails /> */}
       {/* <QuestionAnswerContainer /> */}
-      {/* <RatingReviews /> */}
+      <RatingReviews />
       {/* <RelatedProducts/> */}
     </div>
   );
