+++ client/src/components/RatingReviews/MetaRatings.jsx
@@ -1,0 +1,14 @@
+import React from 'react';
+
+
+const MetaRatings = ({ review }) => {
+  // get request for meta data
+  // calculate average rating from that
+
+
+  return (
+    <div></div>
+  );
+};
+
+export default MetaRatings;
