+++ client/src/components/RatingReviews/RatingsReviews.jsx
@@ -1,0 +1,32 @@
+import React, { useState, useEffect } from 'react';
+import axios from 'axios';
+import MetaRatings from './MetaRatings.jsx';
+import ReviewsList from './ReviewsList';
+
+const RatingReviews = () => {
+  const id = '37311';
+
+  const [reviewList, setReviewList] = useState([]);
+
+  const fetch = () => {
+    axios.get(`/reviews/${id}`, { params: { productID: id } })
+      .then(res => setReviewList(res.data))
+      .catch(err => console.log('failed client get req', err));
+  };
+
+  useEffect(() => {
+    fetch();
+  }, []);
+
+  return (
+    <div>
+      <h1>RATINGS & REVIEWS</h1>
+      <div className='grid grid-cols-3 gap-5'>
+        <div className='col-span-1'><MetaRatings reviewList={reviewList}/></div>
+        <div className='col-span-2'><ReviewsList reviewList={reviewList}/></div>
+      </div>
+    </div>
+  );
+};
+
+export default RatingReviews;
