+++ client/src/components/RatingReviews/ReviewsList.jsx
@@ -1,26 +1,13 @@
-import React, { useState, useEffect } from 'react';
-import axios from 'axios';
+import React, { useState } from 'react';
 import SingleReview from './SingleReview.jsx';
 import NewReview from './NewReview.jsx';
 import moment from 'moment';
 
-const ReviewsList = () => {
+const ReviewsList = ({ reviewList }) => {
   const id = '37311';
 
-  const [reviewList, setReviewList] = useState([]);
   const [visibleReviews, setVisibleReviews] = useState(2);
   const [form, setForm] = useState(false);
-
-  const fetch = () => {
-    axios.get(`/reviews/${id}`, { params: { productID: id } })
-      .then(res => setReviewList(res.data))
-      .catch(err => console.log('failed client get req', err));
-  };
-
-  useEffect(() => {
-    fetch();
-  }, []);
-
 
   const moreClick = () => {
     setVisibleReviews(visibleReviews + 2);
